add chi tiet phieu xuat schema
