+++ server/graphql/schema/phieu-xuat.schema.js
@@ -2,32 +2,18 @@ const { gql } = require('apollo-server');
 const chainMiddlewares = require('../../middlewares');
 const { authRequired } = require('../../middlewares/authentication');
 const phieuXuatController = require('../../controllers/phieu-xuat.controller');
-const chiTietPhieuXuatController = require('../../controllers/chi-tiet-phieu-xuat.controller');
-const { isMongooseModel } = require('../../utils/functions');
-
-const phieuXuat = `
-    id: ID!
-    nguoiXuat: User!
-    createdAt: Date!
-    updatedAt: Date!
-    soMatHangXuat: Int!
-    tongTien: Int!
-`;
 
 const typeDefs = gql`
     type PhieuXuat {
-        ${phieuXuat}
-        chiTiet: [ChiTietPhieuXuat!]!
-    }
-
-    type ChiTietPhieuXuat {
         id: ID!
-        maPhieuXuat: String!
-        sanPham: SanPham
-        soLuongXuat: Int!
-        donGiaXuat: Int!
+        nguoiXuat: User!
+        nguoiMua: String!
+        ngayXuat: Date!
         createdAt: Date!
         updatedAt: Date!
+        soMatHangXuat: Int!
+        tongTien: Int!
+        chiTiet: [ChiTietPhieuXuat!]!
     }
 
     type PhieuXuatsByPage {
@@ -35,10 +21,16 @@ const typeDefs = gql`
         pageInfo: PageInfo!
     }
 
-    input PhieuXuatInput {
+    input ChiTietPhieuXuatInput {
         maSanPham: ID!
         soLuongXuat: Int!
         donGiaXuat: Int!
+        ghiChu: String
+    }
+
+    input UpdatePhieuXuatInput {
+        ngayXuat: Date
+        nguoiMua: String
     }
 
     enum SortPhieuXuat {
@@ -61,69 +53,27 @@ const typeDefs = gql`
         ): PhieuXuat
     }
 
-    type ChiTietPhieuXuatByPage {
-        docs: [ChiTietPhieuXuat!]!
-        pageInfo: PageInfo
-    }
-
-    enum SortChiTietPhieuXuat {
-        NGAY_XUAT_ASC
-        NGAY_XUAT_DESC
-        SO_LUONG_ASC
-        SO_LUONG_DESC
-        DON_GIA_ASC
-        DON_GIA_DESC
-    }
-
-    type ChiTietPhieuXuatQueries {
-        bySanPhamID(
-            id: ID!,
-            page: Int!,
-            limit: Int!,
-            from: Date,
-            to: Date,
-            sort: SortChiTietPhieuXuat
-        ): ChiTietPhieuXuatByPage
-    }
-
     extend type Query {
         phieuXuat: PhieuXuatQueries
-        chiTietPhieuXuat: ChiTietPhieuXuatQueries
     }
 
     type PhieuXuatMutations {
         create(
-            payload: [PhieuXuatInput!]!
+            nguoiMua: String!,
+            ngayXuat: Date!,
+            payload: [ChiTietPhieuXuatInput!]!
         ): PhieuXuat
         delete(
             id: ID!
         ): PhieuXuat
-    }
-
-    type ChiTietPhieuXuatMutationsResponse {
-        phieuXuat: PhieuXuat!
-        sanPhamBiThayDoi: SanPham!
-    }
-
-    type ChiTietPhieuXuatMutations {
-        create(
-            idPhieuXuat: ID!
-            payload: PhieuXuatInput!
+        update (
+            id: ID!,
+            payload: UpdatePhieuXuatInput!
         ): PhieuXuat
-        update(
-            idPhieuXuat: ID!
-            idChiTiet: ID!
-            payload: PhieuXuatInput!
-        ): ChiTietPhieuXuatMutationsResponse
-        delete(
-            idPhieuXuat: ID!
-            idChiTiet: ID!
-        ): ChiTietPhieuXuatMutationsResponse
     }
 
     extend type Mutation {
         phieuXuat: PhieuXuatMutations
-        chiTietPhieuXuat: ChiTietPhieuXuatMutations
     }
 `;
 
@@ -140,29 +90,10 @@ const resolvers = {
     PhieuXuatsByPage: {
         pageInfo: (root) => root
     },
-    SortChiTietPhieuXuat: {
-        NGAY_XUAT_ASC: 'createdAt',
-        NGAY_XUAT_DESC: '-createdAt',
-        SO_LUONG_ASC: 'soLuongXuat',
-        SO_LUONG_DESC: '-soLuongXuat',
-        DON_GIA_ASC: 'donGiaXuat',
-        DON_GIA_DESC: '-donGiaXuat'
-    },
-    ChiTietPhieuXuat: {
-        sanPham: (root) => {
-            return isMongooseModel(root.maSanPham) ? root.maSanPham : null;
-        }
-    },
-    ChiTietPhieuXuatByPage: {
-        pageInfo: (root) => root
-    },
     Query: {
         phieuXuat: chainMiddlewares(authRequired,
             () => ({})
         ),
-        chiTietPhieuXuat: chainMiddlewares(authRequired,
-            () => ({})
-        )
     },
     PhieuXuatQueries: {
         all: async (_, { page, limit, from, to, sort }) =>
@@ -170,36 +101,24 @@ const resolvers = {
         byID: async (_, { id }) =>
             phieuXuatController.getById(id),
     },
-    ChiTietPhieuXuatQueries: {
-        bySanPhamID: async (_, { id, page, limit, from, to, sort }) =>
-            chiTietPhieuXuatController.getBySanPhamID(id, page, limit, from, to, sort)
-    },
     Mutation: {
         phieuXuat: chainMiddlewares(authRequired,
             () => ({})
         ),
-        chiTietPhieuXuat: chainMiddlewares(authRequired,
-            () => ({})
-        ),
     },
     PhieuXuatMutations: {
-        create: async (_, { payload }, { currentUser }) =>
-            phieuXuatController.create(payload, currentUser.id),
-        delete: async (_, { id }) =>
-            phieuXuatController.remove(id)
-    },
-    ChiTietPhieuXuatMutations: {
-        create: async (_, { idPhieuXuat, payload }) =>
-            chiTietPhieuXuatController.create(idPhieuXuat, payload),
-        update: async (_, { idPhieuXuat, idChiTiet, payload }) =>
-            chiTietPhieuXuatController.update(
-                idPhieuXuat,
-                idChiTiet,
+        create: async (_, { nguoiMua, ngayXuat, payload }, { currentUser }) =>
+            phieuXuatController.create(
+                nguoiMua,
+                currentUser.id,
+                ngayXuat,
                 payload,
             ),
-        delete: async (_, { idPhieuXuat, idChiTiet }) =>
-            chiTietPhieuXuatController.remove(idPhieuXuat, idChiTiet),
-    }
+        delete: async (_, { id }, { currentUser }) =>
+            phieuXuatController.remove(id, currentUser),
+        update: async (_, { id, payload }, { currentUser }) =>
+            phieuXuatController.update(id, payload, currentUser),
+    },
 };
 
 module.exports = {
